Flatten the tasty.Reflection cake
Using the cake pattern made it easy to split and keep into smaller files.
But this comes at a cost:
 * Extra bridges
 * Not knowing where to find stuff
 * Hard to update
 * Inability to use opaque types and proper extension methods

The downside is that we will have a larger `Reflection.scala` file, but
it will be proportional to the size of `CompilerInterface.scala` which has
been easier to manage than the cake.

Note: `scala.tasty.Reflection` did not have any API changes.
