--- library/src/scala/tasty/reflect/RootPosition.scala
@@ -1,8 +1,0 @@
-package scala.tasty.reflect
-
-trait RootPosition extends Core {
-
-  /** Root position of this tasty context. For macros it corresponds to the expansion site. */
-  def rootPosition: Position = internal.rootPosition
-
-}
